Address code review: approval race, role-escalation, and Discord bot hardening
Correctness/security:
- Compare self-role-change guard against the DB role, not the JWT, so a
  stale admin token can't re-promote itself (admin user PUT route).
- Claim the approval transition atomically (conditional updateMany gated on
  status) so concurrent approvals can't double-enqueue jobs/notifications.
- Pass requestType to notifications so approved/pending ebook requests render
  ebook-typed embeds (approval service + both request creators).
- Build the live request card from the request's live DB status to close the
  card-ref persistence race.
- Reject empty/negative/non-integer pages in the custom-id decoder.
- Store null (not a synthetic "discord:<id>") as deletedBy when a Discord
  admin-role holder has no linked RMAB account.

Efficiency/cleanup:
- Reuse one REST client across resolveMembersByIds so discord.js throttles
  the batch through a single set of rate-limit buckets.
- Split embeds.ts (over the file-size cap) into embeds/ (book-fields,
  request-cards, approval, lists + barrel) and route list/confirm embeds
  and the request-select menus through shared helpers.

Docs updated for the concurrency-safe approval, the embeds module layout,
and the relationship between the bot approval message and webhook
notifications. Tests updated/added for the atomic claim, ebook notification
typing, the DB-role guard, and the custom-id page guard.
