+++ documentation/integrations/discord-bot.md
@@ -30,7 +30,7 @@ Typed accessor: [src/lib/services/discord/discord-config.ts](../../src/lib/servi
 ## User Mapping
 - `User.discordUserId` (unique, nullable) — set by admins on the Users page edit modal, or in bulk via the Discord tab's **Link Discord Usernames** modal (both `PUT /api/admin/users/[id]`; `role` is optional on that route so a Discord-only update needn't change the role).
 - **Link Discord Usernames modal:** lists all RMAB users; per-row search hits `POST /api/admin/settings/discord/search-members` → `searchGuildMembers` (Discord `GET /guilds/{id}/members/search`, matches username + nickname, filtered to `requester_role_id` when set). Results + existing mappings show as colored avatar pills; selecting one persists the Discord User ID, Unlink clears it.
-- **Existing-mapping pills:** resolved via `POST /api/admin/settings/discord/resolve-members` → `resolveMembersByIds` (guild member → user fallback), cached client-side in `localStorage` ([discordMemberCache.ts](../../src/lib/utils/discordMemberCache.ts), 24h TTL) to avoid re-fetching on every open. The Users-page edit modal shows the same pill right-aligned in the Discord ID field.
+- **Existing-mapping pills:** resolved via `POST /api/admin/settings/discord/resolve-members` → `resolveMembersByIds` (guild member → user fallback; one shared REST client across the batch so discord.js throttles through a single set of rate-limit buckets), cached client-side in `localStorage` ([discordMemberCache.ts](../../src/lib/utils/discordMemberCache.ts), 24h TTL) to avoid re-fetching on every open. The Users-page edit modal shows the same pill right-aligned in the Discord ID field.
 - [discord-user.resolver.ts](../../src/lib/services/discord/discord-user.resolver.ts): `resolveRmabUser(discordUserId)` → `{ user, isAdmin }` (`isAdmin = role==='admin' || isSetupAdmin`); null → ephemeral "not linked" guidance.
 - **Requester gate:** when `requester_role_id` is set, `/request` is allowed only for members holding that role (RMAB admins always pass) — enforced in `handleRequestCommand` via `memberHasRole`.
 
@@ -58,14 +58,15 @@ Typed accessor: [src/lib/services/discord/discord-config.ts](../../src/lib/servi
 
 ## Approval Flow
 - Request needing approval (status `awaiting_approval`) → bot posts to the approval channel: `<@&adminRoleId>` ping + rich embed (title shows **⏳ Pending**; a **Title** field carries the book title with year in parentheses; author, narrator, duration, series #, format, genre, cover — narrator/duration/format audiobook-only, author/narrator top-listed) + **Approve/Deny** buttons. The message location is recorded (`recordApprovalMessage`) so it can be updated on cancel.
-- Click authority: RMAB admin **or** holds `adminRoleId`. Approve/Deny → shared `processRequestApproval` ([request-approval.service.ts](../../src/lib/services/request-approval.service.ts), also used by the Web approve route). Embed title rewritten to **✅ Approved by X** / **🚫 Denied by X**, buttons disabled, the requester's request card refreshed, and the requester DM'd on approval (best-effort).
+- Click authority: RMAB admin **or** holds `adminRoleId`. Approve/Deny → shared `processRequestApproval` ([request-approval.service.ts](../../src/lib/services/request-approval.service.ts), also used by the Web approve route; the transition is claimed atomically so concurrent clicks can't double-process — see [request-approval.md](../admin-features/request-approval.md)). Embed title rewritten to **✅ Approved by X** / **🚫 Denied by X**, buttons disabled, the requester's request card refreshed, and the requester DM'd on approval (best-effort).
+- **Two independent approval surfaces.** The bot's interactive approval message (above, in `admin_notify_channel_id`/`request_channel_id`) is separate from the **Discord webhook notifications** (Settings → Notifications → Discord, [notifications.md](../backend/services/notifications.md)), which also emit a `request_pending_approval`/`request_approved` embed. They serve different setups (the webhook works without the bot; only the bot's message has buttons) and normally reach different audiences (channel announcement vs the requester's DM on approval). **Recommendation:** point the webhook notification channel and the bot approval channel at **different channels** — pointing both at the same channel intentionally produces two messages (one actionable, one informational), not a bug. Suppressing one is avoided because it would regress non-bot/email setups or drop the requester's personal DM.
 
 ## Modules (src/lib/services/discord/)
 - `discord-bot.service.ts` — singleton client; `start/stop/restart/isReady`; registers commands on `ready`; routes interactions.
 - `interaction-router.ts` — dispatch by command name / decoded customId; defers within 3s.
 - `command-definitions.ts` — SlashCommandBuilder defs (guild-scoped JSON).
-- `custom-id.ts` — encode/decode all cross-interaction state into ≤100-char customIds (no server session).
-- `embeds.ts` — embeds, select menus, buttons (severity colors match the notification provider); request-card status footer/color/cancellability helpers.
+- `custom-id.ts` — encode/decode all cross-interaction state into ≤100-char customIds (no server session). Paginated kinds parse the page via `parsePage` (rejects empty/negative/non-integer from tampered IDs → decode returns null).
+- `embeds/` — embed/select/button builders, split to stay under the per-file size cap, re-exported via `embeds/index.ts` (import surface stays `./embeds`): `book-fields.ts` (palette, text helpers, `BookEmbedFields`, shared `addBookFields`, status footer/color/cancellability), `request-cards.ts` (search select, confirm/live request card), `approval.ts` (approval message + decision/cancel rewrites), `lists.ts` (/status & /delete paginated lists, the shared request-select builder, delete-confirm). All list/card field rendering goes through `addBookFields` for one consistent field set.
 - `discord-cards.ts` — post/update the live request card (public + DM); `postRequestCards`, `editRequestCards`.
 - `discord-helpers.ts` — outstanding-status list, actor log context, request queries.
 - `discord-rest.helper.ts` — token-based REST for settings Test/Resolve (no gateway needed).
