+++ src/lib/services/request-creator.service.ts
@@ -309,7 +309,9 @@ export async function createRequestForUser(
     newRequest.id,
     audiobookRecord.title,
     audiobookRecord.author,
-    user.plexUsername || 'Unknown User'
+    user.plexUsername || 'Unknown User',
+    undefined,
+    'audiobook'
   ).catch((error) => {
     logger.error('Failed to queue notification', { error: error instanceof Error ? error.message : String(error) });
   });
